move types and constants to separated files
